+++ gulpfile.cjs
@@ -74,7 +74,6 @@ const sourceFiles = [
   "!Source/ThirdParty/Workers/**",
   "!Source/ThirdParty/google-earth-dbroot-parser.js",
   "!Source/ThirdParty/pako_inflate.js",
-  "!Source/ThirdParty/crunch.js",
 ];
 
 const watchedFiles = [
@@ -126,7 +125,7 @@ function rollupWarning(message) {
   // Ignore eval warnings in third-party code we don't have control over
   if (
     message.code === "EVAL" &&
-    /(protobuf-minimal|crunch)\.js$/.test(message.loc.file)
+    /protobuf-minimal\.js$/.test(message.loc.file)
   ) {
     return;
   }
@@ -737,7 +736,7 @@ function getMimeType(filename) {
     return { type: "text/plain", compress: true };
   } else if (/\.(czml|topojson)$/i.test(filename)) {
     return { type: "application/json", compress: true };
-  } else if (/\.(crn|tgz)$/i.test(filename)) {
+  } else if (/\.tgz$/i.test(filename)) {
     return { type: "application/octet-stream", compress: false };
   }